change test location
