+++ dali/test/python/test_autoserialize.py
